Initial commit.
This commit syncs up the contents of this repository with that of the main
phpMussel repository as it was at commit
c9e582871a3a39bf705edd8eee910b65e6b0ae44, which is shortly after the v2.1.2
release but prior to any subsequent release.

This is in order to provide some continuity with previous phpMussel
releases, while also allowing us to split up that content into several
different repositories, each with a new, fresh commit history.

Note: Deployment from this repository isn't quite recommended just yet. For
now, stick to the main phpMussel, and check back with this repository
later, when it produces its first tagged release. :-)